requirements.txt
# J-Quants API client
jquants-api-client>=1.0.0

# HTTP requests
requests>=2.31.0

# Data analysis
pandas>=2.0.0
numpy>=1.24.0

# Environment variables
python-dotenv>=1.0.0

# Date/Time handling
python-dateutil>=2.8.0

# Logging
loguru>=0.7.0

# Claude API client
anthropic>=0.40.0

# Visualization
matplotlib>=3.7.0

# Web scraping
beautifulsoup4>=4.12.0
